replace doubles to BigDecimal

+++ src/main/java/com/lm/challenge/taxes/controller/dto/output/BasketRSDTO.java
@@ -2,9 +2,11 @@ package com.lm.challenge.taxes.controller.dto.output;
 
 import com.lm.challenge.taxes.controller.dto.base.BasketBaseRDTO;
 
+import java.math.BigDecimal;
+
 public class BasketRSDTO extends BasketBaseRDTO {
 
-    private Double taxes;
+    private BigDecimal taxes;
 
-    private Double totalPrice;
+    private BigDecimal totalPrice;
 }
